requirements.txt
openai
dotenv
python-docx
pypdf
exceptions
argparse
shutil
subprocess
pypdf
python-docx
